Tests/LibWeb/Text/expected/css/CSSPropertyRule-invalid-rules.txt

Number of parsed css rules: 1 (expected: 1)
